refactor: generate-llms-md 改为 vite 插件形式
将 script/generate-llms-md.ts 改造为自包含的 vite 插件（默认导出）：
- dev 模式拦截 /llms/* 实时返回 Markdown / 索引
- build 模式 closeBundle 阶段写入 site/dist/llms

并移除冗余的 site/plugins/llms-md 包装插件，vite.config.ts 直接
从脚本引入插件。
